Update common.js
I noticed that `supportedDevices` is undefined for Mac apps. At the same time, in the current library version, it's not clear if an returned app is for iOS or macOS. This is how I tweaked the library in the project I'm working on now. Perhaps, you'll find a better solution. Just wanted to make you guys aware of the issue.